midterm/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
ecdsa==0.14.1
idna==2.10
pkg-resources==0.0.0
psycopg2==2.8.6
pyasn1==0.4.8
PyJWT==1.7.0
python-jose==3.2.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
